+++ abstract-factory/src/main/java/com/stephanoapiolaza/dessingpattern/abstractfactory/shape/impl/package-info.java
@@ -1,0 +1,8 @@
+/**
+ * Paquete de implementaciones de Figuras Geometricas
+ * 
+ * @version 1.0 - 30-04-2018
+ * @author Stephano.Apiolaza - stephanoapiolaza@gmail.com
+ *
+ */
+package com.stephanoapiolaza.dessingpattern.abstractfactory.shape.impl;