+++ CHANGELOG.md
@@ -1,3 +1,23 @@
+# 0.4.0(30-04-2018)
+
+### Features
+
+* **abstract.factory**: Se agrega patron creacional de objeto.
+
+# 0.3.0(27-01-2018)
+
+### Features
+
+* **prototype**: Se agrega patron creacional de objeto.
+
+# 0.2.0(21-01-2018)
+
+### Features
+
+* **factory.method**: Se agrega patron creacional de clases.
+* **singleton**: Se agrega patron creacional de objetos.
+* **builder**: Se agrega patron creacional de objetos.
+
 # 0.1.0(19-11-2017)
 
 ### Features